drivers: cdi: fix cert_c defect
Add a sanity check to validate the error status of snprintf().
Pass null-terminated string to snprintf(). Passing a character
sequence or wide character sequence that is not null-terminated
can result in accessing memory that is outside the bounds of
the object.

CID 490073
CID 490197
CID 490280
CID 490850

Bug 3512545

Change-Id: I50f10cb0232f9c26fae79efc11131232a5eb03d0
Signed-off-by: Prateek Patel <prpatel@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2707504
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit