nvsciipc: Add configs for build with oot kernel
The OOT kernel uses the mainline kernel and some of the
configs which are used in the nvsciipc are not defined in the
defconfig.

Make the nvsciipc as module if the CONFIG_TEGRA_OOT_MODULE is
enabled.

Bug 3583631

Change-Id: I8c19a0a8bbd5df08ba337546527798272a78bff5
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2713488
(cherry picked from commit 4a844b7fc8ee6a5748552b6777be5d6fd842c1ad)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2728819
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Jonathan Hunter <jonathanh@nvidia.com>
GVS: Gerrit_Virtual_Submit