ASoC: rt5640: Fix the wrong state of JD1 and JD2 on bst1 or bst2 on
When BST1 and BST2 switches are turned ON, JD1 and JD2 voltages drop
to nearly 0.3v. At this level, plug or unplug of jacks on JD1 and JD2
do not generate any interrupt.

This fix is provided by Realtek to address the interrupt issue. With
this the JD1 and JD2 voltages drops are around 0.9v and it is at the
expected level when BST1 and BST2 switches are ON. Interrupts are
verified to work on both JD1 and JD2.

Bug 3645278
Bug 200766032

Change-Id: I95803fd7fcb7622d25f5773f656d428bb98e6d66
Signed-off-by: Oder Chiou <oder_chiou@realtek.com>
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
(cherry picked from commit 51115c1691eaa658033c29347dd30ca067a499ab)
(cherry picked from commit 49869ae90cd7e2f290b744b11a6c513bbad9336f)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2748501
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Mohan Kumar D <mkumard@nvidia.com>