video: tegra: order dc probes in dt sequence
Starting with K5.10, core modules like regulator, i2c, etc. are
being probed late, which has caused DC probes to get deferred.
Because of the deferral, it is no more guaranteed that DC (display
heads) are probed in the order specified in device tree.

This has resulted in bootlogs/prompt (fbconsole) not showing up on
HDMI consistently acorss boots (attached to first head as per DT
sequence).
This affects CLI users and SQA tests that expect HDMI to be primary
display like it was before we moved to K5.10.

To fix this issue, implemented following mechanism to force DC probes
to execute in DT specified order:
1. In dc_common driver, parse and store the sequence of display nodes
   under display-hub node. dc_probe_seq[] array is indexed by
   dc->ctrl_num and holds the number of DCs that are supposed to be
   probed before a given DC's probe can be executed.
2. During dc_probe, check that the number of DCs to be probed before
   current DC are probed using 'probed_dc_count'. If not, defer
   current DC's probe.
3. At the end of dc_probe, increment 'probed_dc_count' to keep count
   of completed DC probes.
4. If a dc probe fails with an error other than -EPROBE_DEFER,
   incremement 'probed_dc_count' so that subsequent DCs can get
   probed.

bug 3639499
bug 200754777

Signed-off-by: Naveen S <nkumars@nvidia.com>
Change-Id: I6c5ee8a1c37be46fd66ec19ab0053560207ac91d
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2741219
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit