thermal: implement kernel driver for t23x OC event
In the past (at least on tegra18x and tegra19x platforms), CCPLEX was
able to learn about overcurrent (OC) events by reading the soctherm
register. But on T23x platforms, reading the value of soctherm register
from CCPLEX is not allowed.

The way to make CCPLEX aware of OC events in tegra23x is to launch an
MRQ to BPMP. So, this driver sends the MRQ to BPMP to provide the
information regarding the overcurrent enable state and the event count
to the userspace via the hwmon sysfs interface. With this driver, the
userspace throttling notification can be published by polling the
associated sysfs nodes.

Bug 3571683

Signed-off-by: yiweiw <yiweiw@nvidia.com>
Change-Id: I6a113f8a44cb326ac0f6e2198f9f1f68e50fc6be
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2748672
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
Reviewed-by: Pekka Jylha-Ollila <pjylhaollila@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>