soc: tegra: cbb: fix master_id retrieval and warn
Fix retrieving the master_id value and further action based on it.
In Tegra194, for the illegal accesses from CCPLEX initiator, we
need to crash the system for fatal errors and warn for non-fatal.
In Tegra234, we need to casue a warning only.

Bug 3622425

Change-Id: I90a3b17ac87882a25cac617783ddcd818faf4b61
Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2751998
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2758319
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>