t234: arvala: set enable-high as default for isolate
PCIE controller will not detect ethernet chip if isolate signal is not
in correct state which is logical high during the probe stage. If there
is no explicit mention of enable-active-high, regulator framework
assumes it to be enable-active-low which will drive isolate signal to
low during probe time; that will cause link to be not detected. This CL
corrects that behavior by setting active high.

Bug 3683634

Signed-off-by: Dipen Patel <dipenp@nvidia.com>
Change-Id: Ifa1f6d1ccdc7497ca9809cfbb701f33cc22787b8
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t23x/common-dts/+/2756815
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit