nvidia: add missing device tree include files
These include files are required to use the nvidia overlay paths
to build out-of-tree device tree binaries. Without them, the in-tree
(non-overlayed) dt-bindings includes must be pulled in. These contain
numerous conflicting include files that result in unpredictable
complilations. Builds as part of the kernel appear to succeed either
due to proximity to the correct files, or slightly different ordering
of include files, but they are likely "getting lucky", as it were,
when pulling in the in-tree dt-bindings directory.

The overlayed paths include everything they need except these two
files, so their omission was likely just an oversight.

Signed-off-by: Kurt Kiefer <kurt.kiefer@arthrex.com>