phy: xusb: Clear CYA_TRK_CODE_UPDATE_ON_IDLE bit
There is an issue of T23x that the logic to drive pad calibrated
values are present in host/device reset domain. Because of this, the
calibration values were getting captured only when the partitions are
out of reset.

This change clears CYA_TRK_CODE_UPDATE_ON_IDLE bit, so it will be
independent of host/device power-gating-partition reset.

Bug 3805352
Bug 4107381

Change-Id: I4679cb224c61437c00a9780d4810518c7bf2f2a3
Signed-off-by: Henry Lin <henryl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2791451
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2907347
Reviewed-by: Wayne Chang <waynec@nvidia.com>
Reviewed-by: Sing-Han Chen <singhanc@nvidia.com>
Reviewed-by: WK Tsai <wtsai@nvidia.com>
Tested-by: Wayne Chang <waynec@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>