cpufreq: tegra194: set offline cores to fmin after sc7 exit
Previously, change was done to set freq of offline cores to Fmin in
commit 8a5e1cb6679 ("driver: cpufreq: put offline cpu cores to fmin").
But during SC7 exit, MCE is updating that Fmin for offline cores to
default value from BCT config file as the MCE doesn't differentiate
between normal boot and SC7 exit. This is resulting in high power
utilization. To fix this, updating the freq of offline cores to Fmin
during SC7 exit (or) system resume. Also, saving the physical_ids and
register offset for all cores during boot to directly access the
register of the offline core. This is done because we need to make smp
call to read the MPIDR register on the target CPU and use that to get
the actmon register offset. But we can't make smp call to read MPIDR
on the offline cpu. The change also helps to reduce the repeated smp
calls on every frequency set request.

Bug 3931770

Change-Id: Ib779b0576ad0056e48635004524b441335900a9b
Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
(cherry picked from commit 8583c3be1043f47b53ae3cd5612916b10f69d6f3)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2866978
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>
Tested-by: Pohsun Su <pohsuns@nvidia.com>