gpio: Driver for Lenovo ThinkEdge SE70 GPIOs
The Lenovo ThinkEdge SE70 provides Digital Input/Output pins connected
internally to a Nuvoton NCT5635Y GPIO expander. Some additional GPIOs are
used to set the protocol for the transceiver IC connected to the serial
ports (which can operate on RS-232, RS-485, RS-422 or LOOP mode).

This commit introduces the driver for this device by abstracting it
through two GPIO controllers: one dedicated to the I/O pins and other
dedicated to set the protocol for the serial ports.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>