crypto: ecc: add fields a and b for ECDH curves
According to SP800-56A section 5.6.2.1, the public key to be processed
for the ECDH operation shall be checked for appropriateness. When the
public key which is considered to be an ephemeral key,
the partial validation test as defined in SP800-56A section
5.6.2.3.4 can be applied.

The partial verification test requires the presence of the field
elements of a and b. For the implemented NIST curves, b is defined in
FIPS 186-4 appendix D.1.2. The element a is implicitly given with the
Weierstrass equation given in D.1.2 where a = p - 3.

Without the fields partial verification test would fail.

This done based on below patch.
https://patchwork.kernel.org/project/linux-crypto/patch/1905067.VJ9nsHHBOt@positron.chronox.de/

Bug 4231422

Change-Id: I779722950ef25ad3ee3badc3ddaef4591042067e
Signed-off-by: Bodla Rakesh Babu <rbodla@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/3006177
Reviewed-by: Nitin Kumbhar <nkumbhar@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>